Struct rustc_binaryen::Module

pub struct Module { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

In-memory representation of a serialized wasm module.

Methods

impl Module

pub fn new(assembly: &CStr, opts: &ModuleOptions) -> Result<Module, String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Creates a new wasm module from the LLVM-assembly provided (in a C string format).

The actual module creation can be tweaked through the various options in ModuleOptions as well. Any errors are just returned as a bland string.

pub fn data(&self) -> &[u8]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the data of the serialized wasm module. This is a foo.wasm file contents.